requirements.txt
semantic-kernel==1.0.3
azure-search-documents
colorama